Diablous and shades
added Diablous, the demon of flight with his often absent wings (damn
shadows). He resides in T-Pose, for you are not worth his time!

Put his model and skeleton with the other models. Make  sure to put his
texture with the others.  His skeleton is called cube.... I know, but
when you convert it from blender, it takes the name of the shape.

toon.material is now using toonTex shaders, so update that file and
place the toonTex files in the GLSL folder.