projectiles, aoes, and Soul!
Fixed AoE and Projectiles.  They were apparently very huge, and had to
be scaled down considerably.
Also added the mSoulNode.  Spells are attached to this instead of
mBodyNode, and mBodyNode itself is attached to mSoulNode. This way, any
changes to scale and pos will only be applied to the body and not the
spells.
Also included is the Toon shader. Let me know if it works kevin, it runs
on Brad's computer and my mac, but not my PC. Cheers!